n64: fix a few memory access opcodes that were using a 32-bit address

+++ ares/n64/cpu/interpreter-fpu.cpp
@@ -554,12 +554,12 @@ auto CPU::FTRUNC_W_D(u8 fd, u8 fs) -> void {
 
 auto CPU::LDC1(u8 ft, cr64& rs, s16 imm) -> void {
   if(!scc.status.enable.coprocessor1) return exception.coprocessor1();
-  if(auto data = read<Dual>(rs.u32 + imm)) FT(u64) = *data;
+  if(auto data = read<Dual>(rs.u64 + imm)) FT(u64) = *data;
 }
 
 auto CPU::LWC1(u8 ft, cr64& rs, s16 imm) -> void {
   if(!scc.status.enable.coprocessor1) return exception.coprocessor1();
-  if(auto data = read<Word>(rs.u32 + imm)) FT(u32) = *data;
+  if(auto data = read<Word>(rs.u64 + imm)) FT(u32) = *data;
 }
 
 auto CPU::MFC1(r64& rt, u8 fs) -> void {
@@ -574,12 +574,12 @@ auto CPU::MTC1(cr64& rt, u8 fs) -> void {
 
 auto CPU::SDC1(u8 ft, cr64& rs, s16 imm) -> void {
   if(!scc.status.enable.coprocessor1) return exception.coprocessor1();
-  write<Dual>(rs.u32 + imm, FT(u64));
+  write<Dual>(rs.u64 + imm, FT(u64));
 }
 
 auto CPU::SWC1(u8 ft, cr64& rs, s16 imm) -> void {
   if(!scc.status.enable.coprocessor1) return exception.coprocessor1();
-  write<Word>(rs.u32 + imm, FT(u32));
+  write<Word>(rs.u64 + imm, FT(u32));
 }
 
 #undef CF
